Fixup InMemoryFilesSystemTests to use #expect(throws:) vs withKnownIssue (#9481)
The `InMemoryFilesSystemTests` were using `withKnownIssue` to validate
that the file system was throwing errors when expected. This was
polluting the test report with tests that appeared to have known issues,
but were actually working as intended.

Use `#expect(throws:)` to validate these functions are throwing the
appropriate error as expected.