Use prebuilts for all targets that build for destination == host.
Add a hidden HostCondition that will check the build parameters
to see if we're building a target for host enable the include and
link settings to use the prebuilts and disable the dependency on
swift-syntax products.

Doesn't support SwiftBuild yet. Need to figure out how to use the
same condition there.

Also cleaned up BuildPlanTests so they are just Native BuildPlan
tests as designed and created a separate BuildTests for the two
tests that used the build system kind setting.