Statically linking Swift Stdlib in SwiftBuild if set (#9417)
Set the `SWIFT_FORCE_STATIC_LINK_STDLIB` build setting in SwiftBuild
accordingly to the command line argument and if it's not on Darwin.

In addition, set the SWIFT_RESOURCE_DIR setting as that is what the
Native build system does.

Fixes: #9320
Issue: rdar://163948614
Depends on: #9389 
Needs: https://github.com/swiftlang/swift-build/pull/942